# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/monitoring/audit/logs": {
      "get": {
        "operationId": "MonitoringController_getAuditLogs",
        "parameters": [
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "审计日志数据"
          }
        },
        "summary": "获取审计日志",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/monitoring/security/events": {
      "get": {
        "operationId": "MonitoringController_getSecurityEvents",
        "parameters": [
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "eventType",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "安全事件数据"
          }
        },
        "summary": "获取安全事件",
        "tags": [
          "monitoring"
        ]
      }
    },
...
    "/api/alerts/rules": {
      "get": {
        "operationId": "AlertController_getAlertRules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功获取告警规则列表"
          }
        },
        "summary": "获取所有告警规则",
        "tags": [
          "告警管理"
        ]
      },
      "post": {
        "operationId": "AlertController_createAlertRule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功创建告警规则"
          },
          "400": {
            "description": "请求参数无效"
          }
        },
        "summary": "创建新的告警规则",
        "tags": [
          "告警管理"
        ]
      }
    },
...
    "/api/alerts/rules/{id}": {
      "put": {
        "operationId": "AlertController_updateAlertRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "告警规则ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功更新告警规则"
          },
          "404": {
            "description": "告警规则不存在"
          }
        },
        "summary": "更新告警规则",
        "tags": [
          "告警管理"
        ]
      },
      "delete": {
        "operationId": "AlertController_deleteAlertRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "告警规则ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功删除告警规则"
          },
          "404": {
            "description": "告警规则不存在"
          }
        },
        "summary": "删除告警规则",
        "tags": [
          "告警管理"
        ]
      }
    },
...
    "/api/alerts/history": {
      "get": {
        "operationId": "AlertController_getAlertHistory",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "告警状态过滤",
            "schema": {}
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "description": "告警级别过滤",
            "schema": {}
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "返回记录数限制",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取告警历史"
          }
        },
        "summary": "获取告警历史",
        "tags": [
          "告警管理"
        ]
      }
    },
    "/api/alerts/{id}/resolve": {
      "post": {
        "operationId": "AlertController_resolveAlert",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "告警ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功解决告警"
          },
          "404": {
            "description": "告警不存在"
          }
        },
        "summary": "手动解决告警",
        "tags": [
          "告警管理"
        ]
      }
    },
...
    "/api/alerts/rules/{id}/test": {
      "post": {
        "operationId": "AlertController_testAlertRule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "告警规则ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功测试告警规则"
          },
          "404": {
            "description": "告警规则不存在"
          }
        },
        "summary": "测试告警规则",
        "tags": [
          "告警管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "Caddy Style Shopping API",
    "description": "\n        ## 电商系统 API 文档\n        \n        ### 功能模块\n        - **认证管理**: 用户注册、登录、JWT令牌管理\n        - **用户管理**: 用户信息、权限管理、地址管理\n        - **商品管理**: 商品CRUD、分类、库存管理\n        - **购物车管理**: 购物车操作、商品选择、批量操作\n        - **订单管理**: 订单创建、支付、状态跟踪\n        - **支付管理**: 多种支付方式、支付回调处理\n        - **系统管理**: 缓存管理、性能监控、健康检查\n        \n        ### 认证方式\n        使用 JWT Bearer Token 进行认证，请在请求头中添加：\n        `Authorization: Bearer <your-token>`\n        \n        ### 响应格式\n        所有API响应都遵循统一格式：\n        ```json\n        {\n          \"code\": 200,\n          \"message\": \"操作成功\",\n          \"data\": {},\n          \"timestamp\": \"2025-01-26T10:30:00Z\",\n          \"requestId\": \"req_123456789\"\n        }\n        ```\n        \n        ### 错误码说明\n        - **1xxx**: 通用错误（参数错误、资源不存在等）\n        - **2xxx**: 认证错误（未授权、令牌过期等）\n        - **3xxx**: 业务错误（库存不足、订单状态错误等）\n        - **5xxx**: 系统错误（服务器错误、数据库错误等）\n      ",
    "version": "1.0.0",
    "contact": {
      "name": "开发团队",
      "url": "https://github.com/your-org/caddy-style-shopping",
      "email": "dev@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    },
    {
      "url": "https://api-staging.example.com",
      "description": "测试环境"
    },
    {
      "url": "https://api.example.com",
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateAlertRuleDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAlertRuleDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}